+++ CHANGELOG.md
@@ -10,7 +10,7 @@
   ([\#5839](https://github.com/cometbft/cometbft/pull/5839))
 - `[mempool]` fix setRecheckFull/setDone race causing spurious ErrRecheckFull.
   ([\#5837](https://github.com/cometbft/cometbft/pull/5837))
-- `[node]` fix(node): close partial listeners on startRPC failure
+- `[node]` close partial listeners on startRPC failure
   ([\#5869](https://github.com/cometbft/cometbft/pull/5869))
 
 ### IMPROVEMENTS